ISL94202: Detect switched off DIS/CHG FETs
The IC might not switch on the FETs even if it is requested. Previously,
the state machine would not detect this. Adding the two additional error
flags BMS_ERR_DIS_OFF and BMS_ERR_CHG_OFF allows to get feedback from
the IC if the FETs are actually on.

The change required to update bms_chg/dis_error functions aswell.