Serial: Use semaphores instead of frequent polling
This allows the serial thread to sleep and avoids polling for incoming
commands every 10 ms as it was done before.

Two semaphores are used, one to lock the receive buffer and the other
one to signal a received command.

Also improving some variable and function naming.

Commit based on charge controller repo.