bq769x2: enable FETs and add custom state machine
The bq769x2 chip handles the FET control autonomously and it is not possible
to switch on the CHG / DSG MOSFETs independently. If not in test mode, the
FETs can be switched off individually, but switched on only all at once.

The common bms_state_machine() function is now defined weak and overridden for
the bq769x2.