Major update of ThingSet Data Object structure for v0.5
The structure has been changed to match the new ThingSet v0.5 spec which
is more suitable to be used in mobile phone apps because the data object
names use a prefix to illustrate if they can be read, written, etc.

All IDs have been kept the same, so previously stored settings in charge
controllers NVM should be compatible.