Remove support for bms_16s100_sc_esp32c3 board
Rationale: This was a very early version of the BMS that later became
the BMS C1. There is no working hardware at Libre Solar anymore.

The variant with STM32G0B1 is still supported. If anyone is still using
the deleted board, please let us know. The BMS C1 v0.3 firmware should
work on that board with minor modifications.